requirements.txt
lmdb==1.2.1
numpy==1.21.3
Pillow==9.5.0
tensorboardX==2.6
torch==1.10.0+cu111
torchvision==0.11.0+cu111
tqdm==4.62.2
